FFWEB-2047: Add UI form in admin panel
