+++ src/Resources/app/administration/src/module/ui-feed-export/page/ui-feed-export-index/ui-feed-export-index.html.twig
@@ -1,7 +1,7 @@
 {% block factfinder_ui_feed_export_index %}
   <sw-page class="ui-feed-export-index">
     <template slot="content">
-      <ui-feed-export-button></ui-feed-export-button>
+      <ui-feed-export-form></ui-feed-export-form>
     </template>
   </sw-page>
 {% endblock %}
