+++ src/Resources/app/administration/src/module/ui-feed-export/snippet/en-GB.json
@@ -1,8 +1,8 @@
 {
   "ui-feed-export": {
-    "title": "Fact-Finder UI feed export",
+    "title": "Fact-Finder feed export",
     "component": {
-      "export-form": {
+      "export_form": {
         "sales_channel": {
           "label": "Sales channel"
         },
@@ -11,6 +11,12 @@
         },
         "button": {
           "title": "Generate feed"
+        },
+        "alert_success": {
+          "text": "Your request is processing"
+        },
+        "alert_error": {
+          "text": "An error occurred during processing Your request"
         }
       }
     }
