+++ src/Resources/public/administration/js/omikron-fact-finder.js
@@ -1,1 +1,1 @@
-(this.webpackJsonp=this.webpackJsonp||[]).push([["omikron-fact-finder"],{"53Qw":function(e,n){e.exports="{% block cms_block_commerce_factfinder_campaigns_preview %}\n  <div>FACT-Finder Campaigns</div>\n{% endblock %}\n\n"},"5K8M":function(e){e.exports=JSON.parse('{"sw-cms":{"blocks":{"commerce":{"factfinderWebComponentsListing":{"label":"FACTFinder Web Components Listing"},"factfinderWebComponentsCampaigns":{"label":"FACTFinder Web Components Campaigns"}}},"elements":{"recordList":{"label":"ff-record-list"},"asn":{"label":"ff-asn"},"sortbox":{"label":"ff-sortbox"},"paging":{"label":"ff-paging"},"campaigns":{"label":"campaigns","config":{"leaveFreeIfNotSpecified":"Leave this value free if you want page to accept all campaigns matching the criteria"}}}}}')},"5LWl":function(e,n){e.exports='{% block sw_cms_element_record_list %}\n  <div class="flex-container">\n    <div>FF-RECORD-LIST</div>\n  </div>\n{% endblock %}\n\n'},"9spW":function(e,n){e.exports='{% block sw_cms_element_asn %}\n  <div class="flex-container">\n    <div>FF-ASN</div>\n  </div>\n{% endblock %}\n\n'},GAqE:function(e,n){e.exports='{% block sw_cms_element_record_list_config %}\n  <div>\n    <sw-switch-field label="subscribe" v-model="element.config.subscribe.value"></sw-switch-field>\n\n    <sw-text-field\n      label="infiniteDebounceDelay"\n      placeholder="Debounce delay"\n      v-model="element.config.infiniteDebounceDelay.value">\n    </sw-text-field>\n\n    <sw-switch-field label="infiniteScrolling" v-model="element.config.infiniteScrolling.value"></sw-switch-field>\n  </div>\n{% endblock %}\n'},MDc6:function(e,n){e.exports="{% block sw_cms_element_campaigns_preview %}\n  <sw-cms-preview-image-text></sw-cms-preview-image-text>\n{% endblock %}\n\n"},MjjZ:function(e){e.exports=JSON.parse('{"ui-feed-export":{"title":"Fact-Finder UI feed export","component":{"export-form":{"sales_channel":{"label":"Sales channel"},"sales_channel_language":{"label":"Language"},"button":{"title":"Generate feed"}}}}}')},O5Wv:function(e,n){e.exports='{% block sw_cms_element_sortbox_preview %}\n  <div class="sw-cms-element-preview-sidebar-sortbox">\n    FF-SORTBOX\n  </div>\n{% endblock %}\n\n'},P45b:function(e,n,t){},Sc8b:function(e,n){e.exports='{% block sw_cms_element_campaigns_config %}\n  <div>\n    <sw-help-text :text="$tc(\'sw-cms.elements.campaigns.config.leaveFreeIfNotSpecified\')">\n    </sw-help-text>\n    <sw-text-field\n      label="Advisor Campaign Name"\n      placeholder="Advisor Campaign Name"\n      v-model="element.config.advisorCampaignName.value">\n    </sw-text-field>\n\n    <sw-help-text :text="$tc(\'sw-cms.elements.campaigns.config.leaveFreeIfNotSpecified\')">\n    </sw-help-text>\n    <sw-text-field\n      label="Feedback Campaign Label"\n      placeholder="Feedback Campaign Label"\n      v-model="element.config.feedbackCampaignLabel.value">\n    </sw-text-field>\n\n  </div>\n{% endblock %}\n'},TQLM:function(e,n){e.exports='{% block cms_block_commerce_factfinder_listing_preview %}\n  <div class="sw-cms-preview-product-three-column">\n    <sw-cms-product-box-preview></sw-cms-product-box-preview>\n    <sw-cms-product-box-preview></sw-cms-product-box-preview>\n    <sw-cms-product-box-preview></sw-cms-product-box-preview>\n  </div>\n{% endblock %}\n\n'},"U+Bf":function(e,n){e.exports='{% block factfinder_ui_feed_export_index %}\n  <sw-page class="ui-feed-export-index">\n    <template slot="content">\n      <ui-feed-export-button></ui-feed-export-button>\n    </template>\n  </sw-page>\n{% endblock %}\n'},Y0es:function(e,n){e.exports='{% block sw_cms_element_campaigns %}\n  <div class="flex-container">\n    <div>CAMPAIGNS</div>\n  </div>\n{% endblock %}\n\n'},"Y4/I":function(e,n,t){"use strict";t.r(n);var s=t("iSSi"),o=t.n(s);const{Component:i}=Shopware;i.register("sw-cms-block-listing",{template:o.a});var c=t("TQLM"),a=t.n(c);const{Component:l}=Shopware;l.register("sw-cms-block-listing-preview",{template:a.a}),Shopware.Service("cmsService").registerCmsBlock({name:"listing",label:"sw-cms.blocks.commerce.factfinderWebComponentsListing.label",category:"commerce",component:"sw-cms-block-listing",previewComponent:"sw-cms-block-listing-preview",slots:{toolbarFilters:"asn",toolbarPaging:"paging",toolbarSorting:"sortbox",records:"record-list"}});var r=t("tmbl"),m=t.n(r);const{Component:p}=Shopware;p.register("sw-cms-block-campaigns",{template:m.a});var d=t("53Qw"),w=t.n(d);const{Component:f}=Shopware;f.register("sw-cms-block-campaigns-preview",{template:w.a}),Shopware.Service("cmsService").registerCmsBlock({name:"campaigns",label:"sw-cms.blocks.commerce.factfinderWebComponentsCampaigns.label",category:"commerce",component:"sw-cms-block-campaigns",previewComponent:"sw-cms-block-campaigns-preview",slots:{campaigns:"campaigns"}});var g=t("5LWl"),b=t.n(g);Shopware.Component.register("sw-cms-el-record-list",{template:b.a,mixins:["cms-element"],created(){this.createdComponent()},methods:{createdComponent(){this.initElementConfig("record-list")}}});var v=t("GAqE"),u=t.n(v);Shopware.Component.register("sw-cms-el-config-record-list",{template:u.a,mixins:["cms-element"],created(){this.createdComponent()},methods:{createdComponent(){this.initElementConfig("record-list")}}});var h=t("cMZj"),x=t.n(h);Shopware.Component.register("sw-cms-el-preview-record-list",{template:x.a}),Shopware.Service("cmsService").registerCmsElement({name:"record-list",label:"sw-cms.elements.recordList.label",component:"sw-cms-el-record-list",configComponent:"sw-cms-el-config-record-list",previewComponent:"sw-cms-el-preview-record-list",defaultConfig:{subscribe:{value:!0,source:"static"},infiniteScrolling:{value:!1,source:"static"},restoreScrollPosition:{value:!1,source:"static"},infiniteDebounceDelay:{value:"32",source:"static"},infiniteScrollMargin:{value:0,source:"static"}}});var C=t("9spW"),_=t.n(C);Shopware.Component.register("sw-cms-el-asn",{template:_.a,mixins:["cms-element"],created(){this.createdComponent()},methods:{createdComponent(){this.initElementConfig("asn")}}});var S=t("rII5"),k=t.n(S);Shopware.Component.register("sw-cms-el-config-asn",{template:k.a,mixins:["cms-element"],created(){this.createdComponent()},methods:{createdComponent(){this.initElementConfig("asn")}}});var F=t("y20k"),y=t.n(F);Shopware.Component.register("sw-cms-el-preview-asn",{template:y.a}),Shopware.Service("cmsService").registerCmsElement({name:"asn",label:"sw-cms.elements.asn.label",component:"sw-cms-el-asn",configComponent:"sw-cms-el-config-asn",previewComponent:"sw-cms-el-preview-asn",defaultConfig:{subscribe:{value:!0,source:"static"}}});var L=t("wMjl"),E=t.n(L);Shopware.Component.register("sw-cms-el-sortbox",{template:E.a,mixins:["cms-element"],created(){this.createdComponent()},methods:{createdComponent(){this.initElementConfig("sortbox")}}});var N=t("wbF8"),O=t.n(N);Shopware.Component.register("sw-cms-el-config-sortbox",{template:O.a,mixins:["cms-element"],created(){this.createdComponent()},methods:{createdComponent(){this.initElementConfig("sortbox")}}});var W=t("O5Wv"),M=t.n(W);Shopware.Component.register("sw-cms-el-preview-sortbox",{template:M.a}),Shopware.Service("cmsService").registerCmsElement({name:"sortbox",label:"sw-cms.elements.sortbox.label",component:"sw-cms-el-sortbox",configComponent:"sw-cms-el-config-sortbox",previewComponent:"sw-cms-el-preview-sortbox",defaultConfig:{subscribe:{value:!0,source:"static"},opened:{value:!0,source:"static"},showSelected:{value:!1,source:"static"},showSelectedFirst:{value:!1,source:"static"},collapseOnblur:{value:!1,source:"static"}}});var A=t("uNAk"),I=t.n(A);Shopware.Component.register("sw-cms-el-paging",{template:I.a,mixins:["cms-element"],created(){this.createdComponent()},methods:{createdComponent(){this.initElementConfig("paging")}}});var T=t("x5H8"),D=t.n(T);Shopware.Component.register("sw-cms-el-config-paging",{template:D.a,mixins:["cms-element"],created(){this.createdComponent()},methods:{createdComponent(){this.initElementConfig("paging")}}});var j=t("ZdP7"),B=t.n(j);Shopware.Component.register("sw-cms-el-preview-paging",{template:B.a}),Shopware.Service("cmsService").registerCmsElement({name:"paging",label:"sw-cms.elements.paging.label",component:"sw-cms-el-paging",configComponent:"sw-cms-el-config-paging",previewComponent:"sw-cms-el-preview-paging",defaultConfig:{subscribe:{value:!0,source:"static"},showOnly:{value:"true",source:"static"}}});var G=t("Y0es"),P=t.n(G);Shopware.Component.register("sw-cms-el-campaigns",{template:P.a,mixins:["cms-element"],created(){this.createdComponent()},methods:{createdComponent(){this.initElementConfig("campaigns")}}});var J=t("Sc8b"),Z=t.n(J);Shopware.Component.register("sw-cms-el-config-campaigns",{template:Z.a,mixins:["cms-element"],created(){this.createdComponent()},methods:{createdComponent(){this.initElementConfig("campaigns")}}});var $=t("MDc6"),K=t.n($);Shopware.Component.register("sw-cms-el-preview-campaigns",{template:K.a}),Shopware.Service("cmsService").registerCmsElement({name:"campaigns",label:"sw-cms.elements.campaigns.label",component:"sw-cms-el-campaigns",configComponent:"sw-cms-el-config-campaigns",previewComponent:"sw-cms-el-preview-campaigns",defaultConfig:{advisorCampaignName:{value:"",source:"static"},feedbackCampaignLabel:{value:"",source:"static"}}});t("lvWF");var Q=t("cjMJ"),R=t.n(Q);const{Component:U}=Shopware;U.register("ui-feed-export-button",{template:R.a,methods:{getFeedExportFile(e){const n=Shopware.Service("syncService").httpClient,t={Authorization:`Bearer ${Shopware.Context.api.authToken.access}`,"Content-Type":"application/json"};n.get(e,{headers:t,params:{}}).then((e=>{alert(e.data)}))}}});var Y=t("U+Bf"),H=t.n(Y);const{Component:q,Mixin:z}=Shopware;q.register("ui-feed-export-index",{template:H.a,mixins:[z.getByName("notification")],metaInfo(){return{title:this.$createTitle()}}});var X=t("fSdK"),V=t("MjjZ");const{Module:ee}=Shopware;ee.register("ui-feed-export",{color:"#ff3d58",icon:"default-shopping-paper-bag-product",title:"ui-feed-export.title",description:"",options:{dataUrl:"test"},snippets:{"de-DE":X,"en-GB":V},routes:{index:{component:"ui-feed-export-index",path:"index"}},navigation:[{label:"ui-feed-export.title",path:"ui.feed.export.index",position:100,parent:"sw-extension"}]});var ne=t("5K8M"),te=t("xOL5");Shopware.Locale.extend("de-DE",ne),Shopware.Locale.extend("en-GB",te)},ZdP7:function(e,n){e.exports="{% block sw_cms_element_paging_preview %}\n  <div>FF-PAGING</div>\n{% endblock %}\n\n"},cMZj:function(e,n){e.exports='{% block sw_cms_element_record_list_preview %}\n  <div class="sw-cms-el-preview-product-box">\n    <sw-cms-product-box-preview :hasText="false"></sw-cms-product-box-preview>\n  </div>\n{% endblock %}\n'},cjMJ:function(e,n){e.exports='<sw-card-view>\n  <sw-card>\n    <sw-container columns="repeat(auto-fit, minmax(250px, 1fr)" gap="0px 30px">\n      <label for="sales_channel">{{ $tc(\'ui-feed-export.component.export-form.sales_channel.label\') }}</label>\n      <sw-entity-single-select entity="sales_channel" id="sales_channel"></sw-entity-single-select>\n\n      <label for="sales_channel_language">{{ $tc(\'ui-feed-export.component.export-form.sales_channel_language.label\') }}</label>\n      <sw-entity-single-select entity="language" id="sales_channel_language"></sw-entity-single-select>\n\n      <sw-button @click="getFeedExportFile(\'_action/fact-finder/generate-feed\')"\n                 variant="success"\n                 block="true"\n                 size="large">\n        {{ $tc(\'ui-feed-export.component.export-form.button.title\') }}\n      </sw-button>\n\n    </sw-container>\n  </sw-card>\n</sw-card-view>\n'},fSdK:function(e){e.exports=JSON.parse('{"factfinder-ui-feed-export":{"helloWorld":"Halo welt"}}')},iSSi:function(e,n){e.exports='{% block cms_block_commerce_factfinder_listing %}\n  <div>\n    <slot name="toolbarFilters"></slot>\n    <div class="flex-container">\n      <slot name="toolbarPaging"></slot>\n      <slot name="toolbarSorting"></slot>\n    </div>\n    <slot name="records"></slot>\n  </div>\n{% endblock %}\n'},lvWF:function(e,n,t){var s=t("P45b");"string"==typeof s&&(s=[[e.i,s,""]]),s.locals&&(e.exports=s.locals);(0,t("SZ7m").default)("17c245ee",s,!0,{})},rII5:function(e,n){e.exports='{% block sw_cms_element_asn_config %}\n  <div>\n    <sw-switch-field label="subscribe" v-model="element.config.subscribe.value"></sw-switch-field>\n  </div>\n{% endblock %}\n'},tmbl:function(e,n){e.exports='{% block cms_block_commerce_factfinder_campaigns %}\n      <div class="flex-container">\n        <slot name="campaigns"></slot>\n      </div>\n{% endblock %}\n'},uNAk:function(e,n){e.exports='{% block sw_cms_element_paging %}\n  <div class="flex-container">\n    <div>FF-PAGING</div>\n  </div>\n{% endblock %}\n\n'},wMjl:function(e,n){e.exports='{% block sw_cms_element_sortbox %}\n  <div class="flex-container">\n    <div>FF-SORTBOX</div>\n  </div>\n{% endblock %}\n\n'},wbF8:function(e,n){e.exports='{% block sw_cms_element_sortbox_config %}\n  <div>\n    <sw-switch-field label="subscribe" v-model="element.config.subscribe.value"></sw-switch-field>\n    <sw-switch-field label="opened" v-model="element.config.showSelected.value"></sw-switch-field>\n    <sw-switch-field label="show selected" v-model="element.config.showSelectedFirst.value"></sw-switch-field>\n    <sw-switch-field label="show selected first" v-if="element.config.showSelected.value" v-model="element.config.showSelectedFirst.value"></sw-switch-field>\n    <sw-switch-field label="collapse onblur" v-model="element.config.collapseOnblur.value"></sw-switch-field>\n  </div>\n{% endblock %}\n'},x5H8:function(e,n){e.exports='{% block sw_cms_element_paging_config %}\n  <div>\n    <sw-text-field\n      label="Show Only"\n      placeholder="Show Only"\n      v-model="element.config.showOnly.value">\n    </sw-text-field>\n    <sw-switch-field label="Subscribe" v-model="element.config.subscribe.value"></sw-switch-field>\n  </div>\n{% endblock %}\n'},xOL5:function(e){e.exports=JSON.parse('{"sw-cms":{"blocks":{"commerce":{"factfinderWebComponentsListing":{"label":"FACTFinder Web Components Listing"},"factfinderWebComponentsCampaigns":{"label":"FACTFinder Web Components Campaigns"}}},"elements":{"recordList":{"label":"ff-record-list"},"asn":{"label":"ff--asn"},"sortbox":{"label":"ff-sortbox"},"paging":{"label":"ff-paging"},"campaigns":{"label":"campaigns","config":{"leaveFreeIfNotSpecified":"Leave this value free if you want page to accept all campaigns matching the criteria"}}}}}')},y20k:function(e,n){e.exports='{% block sw_cms_preview_element_asn %}\n  <div class="sw-cms-element-preview-sidebar-filter">\n    <sw-cms-el-preview-sidebar-filter></sw-cms-el-preview-sidebar-filter>\n  </div>\n{% endblock %}\n\n'}},[["Y4/I","runtime","vendors-node"]]]);
+(this.webpackJsonp=this.webpackJsonp||[]).push([["omikron-fact-finder"],{"53Qw":function(e,n){e.exports="{% block cms_block_commerce_factfinder_campaigns_preview %}\n  <div>FACT-Finder Campaigns</div>\n{% endblock %}\n\n"},"5K8M":function(e){e.exports=JSON.parse('{"sw-cms":{"blocks":{"commerce":{"factfinderWebComponentsListing":{"label":"FACTFinder Web Components Listing"},"factfinderWebComponentsCampaigns":{"label":"FACTFinder Web Components Campaigns"}}},"elements":{"recordList":{"label":"ff-record-list"},"asn":{"label":"ff-asn"},"sortbox":{"label":"ff-sortbox"},"paging":{"label":"ff-paging"},"campaigns":{"label":"campaigns","config":{"leaveFreeIfNotSpecified":"Leave this value free if you want page to accept all campaigns matching the criteria"}}}}}')},"5LWl":function(e,n){e.exports='{% block sw_cms_element_record_list %}\n  <div class="flex-container">\n    <div>FF-RECORD-LIST</div>\n  </div>\n{% endblock %}\n\n'},"9spW":function(e,n){e.exports='{% block sw_cms_element_asn %}\n  <div class="flex-container">\n    <div>FF-ASN</div>\n  </div>\n{% endblock %}\n\n'},GAqE:function(e,n){e.exports='{% block sw_cms_element_record_list_config %}\n  <div>\n    <sw-switch-field label="subscribe" v-model="element.config.subscribe.value"></sw-switch-field>\n\n    <sw-text-field\n      label="infiniteDebounceDelay"\n      placeholder="Debounce delay"\n      v-model="element.config.infiniteDebounceDelay.value">\n    </sw-text-field>\n\n    <sw-switch-field label="infiniteScrolling" v-model="element.config.infiniteScrolling.value"></sw-switch-field>\n  </div>\n{% endblock %}\n'},M0Qd:function(e,n){e.exports='<sw-card-view>\n\n  <sw-card>\n\n    <sw-alert\n      v-if="successAlertVisible"\n      variant="success"\n      appearance="notification"\n      :showIcon="true"\n      :closable="true">\n      {{ $tc(\'ui-feed-export.component.export_form.alert_success.text\') }}\n    </sw-alert>\n\n    <sw-alert\n      v-if="errorAlertVisible"\n      variant="warning"\n      appearance="notification"\n      :showIcon="true"\n      :closable="true">\n      {{ $tc(\'ui-feed-export.component.export_form.alert_error.text\') }}\n    </sw-alert>\n\n    <sw-container columns="repeat(auto-fit, minmax(250px, 1fr)" gap="0px 30px">\n      <sw-entity-single-select\n        v-model="salesChannelValue"\n        entity="sales_channel"\n        id="sales_channel"\n        :label="$tc(\'ui-feed-export.component.export_form.sales_channel.label\')">\n      </sw-entity-single-select>\n\n      <sw-entity-single-select\n        v-model="salesChannelLanguageValue"\n        entity="language"\n        id="sales_channel_language"\n        :label="$tc(\'ui-feed-export.component.export_form.sales_channel_language.label\')">\n      </sw-entity-single-select>\n\n      <sw-button @click="getFeedExportFile(\'_action/fact-finder/generate-feed\')"\n                 variant="success"\n                 block="true"\n                 size="large">\n        {{ $tc(\'ui-feed-export.component.export_form.button.title\') }}\n      </sw-button>\n\n    </sw-container>\n  </sw-card>\n</sw-card-view>\n'},MDc6:function(e,n){e.exports="{% block sw_cms_element_campaigns_preview %}\n  <sw-cms-preview-image-text></sw-cms-preview-image-text>\n{% endblock %}\n\n"},MjjZ:function(e){e.exports=JSON.parse('{"ui-feed-export":{"title":"Fact-Finder feed export","component":{"export_form":{"sales_channel":{"label":"Sales channel"},"sales_channel_language":{"label":"Language"},"button":{"title":"Generate feed"},"alert_success":{"text":"Your request is processing"},"alert_error":{"text":"An error occurred during processing Your request"}}}}}')},O5Wv:function(e,n){e.exports='{% block sw_cms_element_sortbox_preview %}\n  <div class="sw-cms-element-preview-sidebar-sortbox">\n    FF-SORTBOX\n  </div>\n{% endblock %}\n\n'},P45b:function(e,n,t){},Sc8b:function(e,n){e.exports='{% block sw_cms_element_campaigns_config %}\n  <div>\n    <sw-help-text :text="$tc(\'sw-cms.elements.campaigns.config.leaveFreeIfNotSpecified\')">\n    </sw-help-text>\n    <sw-text-field\n      label="Advisor Campaign Name"\n      placeholder="Advisor Campaign Name"\n      v-model="element.config.advisorCampaignName.value">\n    </sw-text-field>\n\n    <sw-help-text :text="$tc(\'sw-cms.elements.campaigns.config.leaveFreeIfNotSpecified\')">\n    </sw-help-text>\n    <sw-text-field\n      label="Feedback Campaign Label"\n      placeholder="Feedback Campaign Label"\n      v-model="element.config.feedbackCampaignLabel.value">\n    </sw-text-field>\n\n  </div>\n{% endblock %}\n'},TQLM:function(e,n){e.exports='{% block cms_block_commerce_factfinder_listing_preview %}\n  <div class="sw-cms-preview-product-three-column">\n    <sw-cms-product-box-preview></sw-cms-product-box-preview>\n    <sw-cms-product-box-preview></sw-cms-product-box-preview>\n    <sw-cms-product-box-preview></sw-cms-product-box-preview>\n  </div>\n{% endblock %}\n\n'},"U+Bf":function(e,n){e.exports='{% block factfinder_ui_feed_export_index %}\n  <sw-page class="ui-feed-export-index">\n    <template slot="content">\n      <ui-feed-export-form></ui-feed-export-form>\n    </template>\n  </sw-page>\n{% endblock %}\n'},Y0es:function(e,n){e.exports='{% block sw_cms_element_campaigns %}\n  <div class="flex-container">\n    <div>CAMPAIGNS</div>\n  </div>\n{% endblock %}\n\n'},"Y4/I":function(e,n,t){"use strict";t.r(n);var s=t("iSSi"),o=t.n(s);const{Component:i}=Shopware;i.register("sw-cms-block-listing",{template:o.a});var a=t("TQLM"),l=t.n(a);const{Component:c}=Shopware;c.register("sw-cms-block-listing-preview",{template:l.a}),Shopware.Service("cmsService").registerCmsBlock({name:"listing",label:"sw-cms.blocks.commerce.factfinderWebComponentsListing.label",category:"commerce",component:"sw-cms-block-listing",previewComponent:"sw-cms-block-listing-preview",slots:{toolbarFilters:"asn",toolbarPaging:"paging",toolbarSorting:"sortbox",records:"record-list"}});var r=t("tmbl"),m=t.n(r);const{Component:p}=Shopware;p.register("sw-cms-block-campaigns",{template:m.a});var d=t("53Qw"),w=t.n(d);const{Component:f}=Shopware;f.register("sw-cms-block-campaigns-preview",{template:w.a}),Shopware.Service("cmsService").registerCmsBlock({name:"campaigns",label:"sw-cms.blocks.commerce.factfinderWebComponentsCampaigns.label",category:"commerce",component:"sw-cms-block-campaigns",previewComponent:"sw-cms-block-campaigns-preview",slots:{campaigns:"campaigns"}});var g=t("5LWl"),b=t.n(g);Shopware.Component.register("sw-cms-el-record-list",{template:b.a,mixins:["cms-element"],created(){this.createdComponent()},methods:{createdComponent(){this.initElementConfig("record-list")}}});var v=t("GAqE"),u=t.n(v);Shopware.Component.register("sw-cms-el-config-record-list",{template:u.a,mixins:["cms-element"],created(){this.createdComponent()},methods:{createdComponent(){this.initElementConfig("record-list")}}});var h=t("cMZj"),x=t.n(h);Shopware.Component.register("sw-cms-el-preview-record-list",{template:x.a}),Shopware.Service("cmsService").registerCmsElement({name:"record-list",label:"sw-cms.elements.recordList.label",component:"sw-cms-el-record-list",configComponent:"sw-cms-el-config-record-list",previewComponent:"sw-cms-el-preview-record-list",defaultConfig:{subscribe:{value:!0,source:"static"},infiniteScrolling:{value:!1,source:"static"},restoreScrollPosition:{value:!1,source:"static"},infiniteDebounceDelay:{value:"32",source:"static"},infiniteScrollMargin:{value:0,source:"static"}}});var C=t("9spW"),_=t.n(C);Shopware.Component.register("sw-cms-el-asn",{template:_.a,mixins:["cms-element"],created(){this.createdComponent()},methods:{createdComponent(){this.initElementConfig("asn")}}});var S=t("rII5"),k=t.n(S);Shopware.Component.register("sw-cms-el-config-asn",{template:k.a,mixins:["cms-element"],created(){this.createdComponent()},methods:{createdComponent(){this.initElementConfig("asn")}}});var F=t("y20k"),L=t.n(F);Shopware.Component.register("sw-cms-el-preview-asn",{template:L.a}),Shopware.Service("cmsService").registerCmsElement({name:"asn",label:"sw-cms.elements.asn.label",component:"sw-cms-el-asn",configComponent:"sw-cms-el-config-asn",previewComponent:"sw-cms-el-preview-asn",defaultConfig:{subscribe:{value:!0,source:"static"}}});var y=t("wMjl"),A=t.n(y);Shopware.Component.register("sw-cms-el-sortbox",{template:A.a,mixins:["cms-element"],created(){this.createdComponent()},methods:{createdComponent(){this.initElementConfig("sortbox")}}});var E=t("wbF8"),N=t.n(E);Shopware.Component.register("sw-cms-el-config-sortbox",{template:N.a,mixins:["cms-element"],created(){this.createdComponent()},methods:{createdComponent(){this.initElementConfig("sortbox")}}});var O=t("O5Wv"),M=t.n(O);Shopware.Component.register("sw-cms-el-preview-sortbox",{template:M.a}),Shopware.Service("cmsService").registerCmsElement({name:"sortbox",label:"sw-cms.elements.sortbox.label",component:"sw-cms-el-sortbox",configComponent:"sw-cms-el-config-sortbox",previewComponent:"sw-cms-el-preview-sortbox",defaultConfig:{subscribe:{value:!0,source:"static"},opened:{value:!0,source:"static"},showSelected:{value:!1,source:"static"},showSelectedFirst:{value:!1,source:"static"},collapseOnblur:{value:!1,source:"static"}}});var W=t("uNAk"),I=t.n(W);Shopware.Component.register("sw-cms-el-paging",{template:I.a,mixins:["cms-element"],created(){this.createdComponent()},methods:{createdComponent(){this.initElementConfig("paging")}}});var T=t("x5H8"),V=t.n(T);Shopware.Component.register("sw-cms-el-config-paging",{template:V.a,mixins:["cms-element"],created(){this.createdComponent()},methods:{createdComponent(){this.initElementConfig("paging")}}});var D=t("ZdP7"),G=t.n(D);Shopware.Component.register("sw-cms-el-preview-paging",{template:G.a}),Shopware.Service("cmsService").registerCmsElement({name:"paging",label:"sw-cms.elements.paging.label",component:"sw-cms-el-paging",configComponent:"sw-cms-el-config-paging",previewComponent:"sw-cms-el-preview-paging",defaultConfig:{subscribe:{value:!0,source:"static"},showOnly:{value:"true",source:"static"}}});var B=t("Y0es"),P=t.n(B);Shopware.Component.register("sw-cms-el-campaigns",{template:P.a,mixins:["cms-element"],created(){this.createdComponent()},methods:{createdComponent(){this.initElementConfig("campaigns")}}});var j=t("Sc8b"),$=t.n(j);Shopware.Component.register("sw-cms-el-config-campaigns",{template:$.a,mixins:["cms-element"],created(){this.createdComponent()},methods:{createdComponent(){this.initElementConfig("campaigns")}}});var Y=t("MDc6"),Z=t.n(Y);Shopware.Component.register("sw-cms-el-preview-campaigns",{template:Z.a}),Shopware.Service("cmsService").registerCmsElement({name:"campaigns",label:"sw-cms.elements.campaigns.label",component:"sw-cms-el-campaigns",configComponent:"sw-cms-el-config-campaigns",previewComponent:"sw-cms-el-preview-campaigns",defaultConfig:{advisorCampaignName:{value:"",source:"static"},feedbackCampaignLabel:{value:"",source:"static"}}});t("lvWF");var J=t("M0Qd"),Q=t.n(J);const{Component:q}=Shopware;q.register("ui-feed-export-form",{template:Q.a,data:()=>({salesChannelValue:null,salesChannelLanguageValue:null,successAlertVisible:!1,errorAlertVisible:!1}),methods:{getFeedExportFile(e){const n=Shopware.Service("syncService").httpClient,t={Authorization:`Bearer ${Shopware.Context.api.authToken.access}`,"Content-Type":"application/json"},s={salesChannelValue:this.salesChannelValue,salesChannelLanguageValue:this.salesChannelLanguageValue};n.get(e,{headers:t,params:s}).then((e=>{200===e.status?this.successAlertVisible=!0:this.errorAlertVisible=!0}))}}});var K=t("U+Bf"),R=t.n(K);const{Component:U,Mixin:z}=Shopware;U.register("ui-feed-export-index",{template:R.a,mixins:[z.getByName("notification")],metaInfo(){return{title:this.$createTitle()}}});var H=t("fSdK"),X=t("MjjZ");const{Module:ee}=Shopware;ee.register("ui-feed-export",{color:"#ff3d58",icon:"default-shopping-paper-bag-product",title:"ui-feed-export.title",description:"",options:{dataUrl:"test"},snippets:{"de-DE":H,"en-GB":X},routes:{index:{component:"ui-feed-export-index",path:"index"}},navigation:[{label:"ui-feed-export.title",path:"ui.feed.export.index",position:100,parent:"sw-extension"}]});var ne=t("5K8M"),te=t("xOL5");Shopware.Locale.extend("de-DE",ne),Shopware.Locale.extend("en-GB",te)},ZdP7:function(e,n){e.exports="{% block sw_cms_element_paging_preview %}\n  <div>FF-PAGING</div>\n{% endblock %}\n\n"},cMZj:function(e,n){e.exports='{% block sw_cms_element_record_list_preview %}\n  <div class="sw-cms-el-preview-product-box">\n    <sw-cms-product-box-preview :hasText="false"></sw-cms-product-box-preview>\n  </div>\n{% endblock %}\n'},fSdK:function(e){e.exports=JSON.parse('{"ui-feed-export":{"title":"Fact-Finder feed export","component":{"export_form":{"sales_channel":{"label":"Sales channel"},"sales_channel_language":{"label":"Language"},"button":{"title":"Generate feed"},"alert_success":{"text":"Your request is processing"}}}}}')},iSSi:function(e,n){e.exports='{% block cms_block_commerce_factfinder_listing %}\n  <div>\n    <slot name="toolbarFilters"></slot>\n    <div class="flex-container">\n      <slot name="toolbarPaging"></slot>\n      <slot name="toolbarSorting"></slot>\n    </div>\n    <slot name="records"></slot>\n  </div>\n{% endblock %}\n'},lvWF:function(e,n,t){var s=t("P45b");"string"==typeof s&&(s=[[e.i,s,""]]),s.locals&&(e.exports=s.locals);(0,t("SZ7m").default)("17c245ee",s,!0,{})},rII5:function(e,n){e.exports='{% block sw_cms_element_asn_config %}\n  <div>\n    <sw-switch-field label="subscribe" v-model="element.config.subscribe.value"></sw-switch-field>\n  </div>\n{% endblock %}\n'},tmbl:function(e,n){e.exports='{% block cms_block_commerce_factfinder_campaigns %}\n      <div class="flex-container">\n        <slot name="campaigns"></slot>\n      </div>\n{% endblock %}\n'},uNAk:function(e,n){e.exports='{% block sw_cms_element_paging %}\n  <div class="flex-container">\n    <div>FF-PAGING</div>\n  </div>\n{% endblock %}\n\n'},wMjl:function(e,n){e.exports='{% block sw_cms_element_sortbox %}\n  <div class="flex-container">\n    <div>FF-SORTBOX</div>\n  </div>\n{% endblock %}\n\n'},wbF8:function(e,n){e.exports='{% block sw_cms_element_sortbox_config %}\n  <div>\n    <sw-switch-field label="subscribe" v-model="element.config.subscribe.value"></sw-switch-field>\n    <sw-switch-field label="opened" v-model="element.config.showSelected.value"></sw-switch-field>\n    <sw-switch-field label="show selected" v-model="element.config.showSelectedFirst.value"></sw-switch-field>\n    <sw-switch-field label="show selected first" v-if="element.config.showSelected.value" v-model="element.config.showSelectedFirst.value"></sw-switch-field>\n    <sw-switch-field label="collapse onblur" v-model="element.config.collapseOnblur.value"></sw-switch-field>\n  </div>\n{% endblock %}\n'},x5H8:function(e,n){e.exports='{% block sw_cms_element_paging_config %}\n  <div>\n    <sw-text-field\n      label="Show Only"\n      placeholder="Show Only"\n      v-model="element.config.showOnly.value">\n    </sw-text-field>\n    <sw-switch-field label="Subscribe" v-model="element.config.subscribe.value"></sw-switch-field>\n  </div>\n{% endblock %}\n'},xOL5:function(e){e.exports=JSON.parse('{"sw-cms":{"blocks":{"commerce":{"factfinderWebComponentsListing":{"label":"FACTFinder Web Components Listing"},"factfinderWebComponentsCampaigns":{"label":"FACTFinder Web Components Campaigns"}}},"elements":{"recordList":{"label":"ff-record-list"},"asn":{"label":"ff--asn"},"sortbox":{"label":"ff-sortbox"},"paging":{"label":"ff-paging"},"campaigns":{"label":"campaigns","config":{"leaveFreeIfNotSpecified":"Leave this value free if you want page to accept all campaigns matching the criteria"}}}}}')},y20k:function(e,n){e.exports='{% block sw_cms_preview_element_asn %}\n  <div class="sw-cms-element-preview-sidebar-filter">\n    <sw-cms-el-preview-sidebar-filter></sw-cms-el-preview-sidebar-filter>\n  </div>\n{% endblock %}\n\n'}},[["Y4/I","runtime","vendors-node"]]]);
